+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - (Bugfix) (Platform) Cover NoAuth Case for Identity Service
 - (Feature) Add ArangoMember Args
 - (Maintenance) Unify References via golangci-linter
+- (Feature) Add Default Container Mods
 
 ## [1.2.47](https://github.com/arangodb/kube-arangodb/tree/1.2.47) (2025-03-28)
 - (Bugfix) Use Profile Annotations